SOLR-15960: Move EnvUtils to Solrj common (#2192)
Also use EncUtils for property substitution in solr.xml
