+++ solr/core/src/java/org/apache/solr/util/StartupLoggingUtils.java
@@ -27,6 +27,7 @@ import org.apache.logging.log4j.core.appender.AbstractOutputStreamAppender;
 import org.apache.logging.log4j.core.appender.ConsoleAppender;
 import org.apache.logging.log4j.core.config.Configuration;
 import org.apache.logging.log4j.core.config.LoggerConfig;
+import org.apache.solr.common.util.EnvUtils;
 import org.apache.solr.common.util.SuppressForbidden;
 import org.slf4j.ILoggerFactory;
 import org.slf4j.Logger;
