+++ solr/solrj/src/test/org/apache/solr/common/util/EnvUtilsTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.solr.util;
+package org.apache.solr.common.util;
 
 import java.util.List;
 import java.util.Map;
